Fix: 뷰에 필요한 뷰모델만 불러오게 정리
